Add reference to upstream's problem with Pango source code in the changelog.
Pango is used for building the GTK+ runtime in Windows, but it is a custom build in upstream, for which no source code is currently available. However Pidgin++ since revision 72 uses a newer version from Fedora, for which the source code is indeed available, and since revision 74 it can build the source version of the GTK+ runtime for distribution alongside the binary bundle. The problem with upstream's Pango has been reported here: https://developer.pidgin.im/ticket/16285.