Workaround for the slow GTK+ extraction in Windows installer.
The extraction of the GTK+ downloaded by the installer was performed by a separate NSIS script that would extract each file individually, except for the locale files. However a recent GTK+ version that includes a share/mime folder started to make this very slow. The GTK+ bundle is now completely extracted, then the share/locale folder is deleted, then the locales selected in the installer are extracted individually once again. The result is the same as before, as intended by revision 64, but is much faster.