URI handler for the IRC protocol.
A new temporary account API is now available and is used by the IRC protocol for
URIs that cannot be handled by any connected account. Bouncer accounts are
supported as long as they specify the original server in the bounced_domain
preference.

The signal API now allows connecting to a signal without duplication. This is
used for deleting temporary accounts when all related conversations are closed:
the callback is registered only once when a given URI is handled multiple times.